+++ ueb01/ctime.hpp
@@ -5,11 +5,11 @@ class CTime {
 
 public:
     CTime();
-    CTime(int hours, int minutes, int seconds = 0);
+    CTime(short hours, short minutes, short seconds = 0);
     void print();
 
 private:
-    int m_hours, m_minutes, m_seconds;
+    short m_hours, m_minutes, m_seconds;
 };
